:truck: Move ToolbarItemDragger to toolbar directory
